Refactor labeling state management into a separate LabelingState class
- Moved the LabelingState class from labeling-panel.js to a new LabelingState.js file for better organization and maintainability.
- Updated labeling-panel.js to import the LabelingState class, streamlining the code structure.
- Enhanced state management functionality, including undo/redo capabilities and label interaction methods.
